third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-targetY-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'targetY' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.targetY.baseVal is 2
PASS convolveMatrixElement.targetY.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
